Added Register Modal
I added the Register Modal and added register to the Redux. However, there is a risk, in that the mapping berween items and a new user doesn't occur correctly. In addition, the Login methods in the reducer need to be changed as to ensure that that the items of a user and that user have to occur. Since, @jarhun88  worked on the server side. I think it would be easier for him to do the disaptch and ensure that none of the code breaks.